Move adjusted_now to base Scheduler class
This is a public accessed function, and the implementation is not
specific to SM2
