+++ mnemosyne/libmnemosyne/schedulers/SM2_mnemosyne.py
@@ -35,36 +35,6 @@ class SM2Mnemosyne(Scheduler):
     name = "SM2 Mnemosyne"
     warned_about_too_many_cards = False  # default false
 
-    def adjusted_now(self, now=None):
-
-        """Timezone information and 'day_starts_at' will only become relevant
-        when the queue is built, not at schedule time, to allow for
-        moving to a different timezone after a card has been scheduled.
-        Cards are due when 'adjusted_now >= next_rep', and this function
-        makes sure that happens at h:00 local time (with h being
-        'day_starts_at').
-
-        """
-
-        if now == None:
-            now = time.time()
-        # The larger 'day_starts_at', the later the card should become due,
-        # i.e. larger than 'next_card', so the more 'now' should be decreased.
-        now -= self.config()["day_starts_at"] * HOUR
-        # 'altzone' or 'timezone' contains the offset in seconds west of UTC.
-        # This number is positive for the US, where a card should become
-        # due later than in Europe, so 'now' should be decreased by this
-        # offset.
-        # As for when to use 'altzone' instead of 'timezone' if daylight
-        # savings time is active, this is a matter of big confusion
-        # among the Python developers themselves:
-        # http://bugs.python.org/issue7229
-        if time.localtime(now).tm_isdst and time.daylight:
-            now -= time.altzone
-        else:
-            now -= time.timezone
-        return int(now)
-
     def true_scheduled_interval(self, card):
 
         """Since 'next_rep' is always midnight UTC for retention reps, we need
